Attempted to fix better Rubocop. Did so, but also tutoring spec is broken. Not my fault, sign in is required now.

+++ spec/feature/student_scheduler_spec.rb
@@ -1,40 +1,36 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 RSpec.describe 'Student scheduler', type: :feature do
   describe 'student scheduler' do
-
     it 'schedules a tutoring session' do
       # Create some data
-      user1 = User.create!(:first_name => "John",
-                   :last_name => "Doe",
-                   :email => "john@tamu.edu",
-                   :password => "T3st!!b"
-      )
+      user1 = User.create!(first_name: 'John',
+                           last_name: 'Doe',
+                           email: 'john@tamu.edu',
+                           password: 'T3st!!b')
 
-      user2 = User.create!(:first_name => "Jane",
-                   :last_name => "Doe",
-                   :email => "jane@tamu.edu",
-                   :password => "T3st!!c"
-      )
+      user2 = User.create!(first_name: 'Jane',
+                           last_name: 'Doe',
+                           email: 'jane@tamu.edu',
+                           password: 'T3st!!c')
 
-      user3 = User.create!(:first_name => "Jeff",
-                   :last_name => "Doe",
-                   :email => "jeff@tamu.edu",
-                   :password => "T3st!!d"
-      )
+      user3 = User.create!(first_name: 'Jeff',
+                           last_name: 'Doe',
+                           email: 'jeff@tamu.edu',
+                           password: 'T3st!!d')
 
-      tsession1 = TutoringSession.create!(:id => 1,
-                             :tutor_id => user2.id,
-                             :scheduled_datetime => Time.now,
-                             :completed_datetime => nil,
-                             :session_status => ""
-      )
+      tsession1 = TutoringSession.create!(id: 1,
+                                          tutor_id: user2.id,
+                                          scheduled_datetime: Time.now,
+                                          completed_datetime: nil,
+                                          session_status: '')
 
-      tsession2 = TutoringSession.create!(:id => 2,
-                             :tutor_id => user3.id,
-                             :scheduled_datetime => Time.now,
-                             :completed_datetime => nil,
-                             :session_status => ""
-      )
+      tsession2 = TutoringSession.create!(id: 2,
+                                          tutor_id: user3.id,
+                                          scheduled_datetime: Time.now,
+                                          completed_datetime: nil,
+                                          session_status: '')
 
       # Start at student index page
       visit 'student/index'
@@ -65,18 +61,14 @@ RSpec.describe 'Student scheduler', type: :feature do
       # Make sure join table worked
       @count = false
       TutoringSession.find(tsession1.id).users.each do |user|
-        if user.id == user1.id
-          @count = true
-        end
+        @count = true if user.id == user1.id
       end
       expect(@count).to eq(true)
 
       # Delete everything
       sessions = TutoringSession.all
       sessions.each do |session|
-        unless session.users.blank?
-          session.users.destroy_all
-        end
+        session.users.destroy_all unless session.users.blank?
       end
       TutoringSession.delete_all
       User.delete_all
